src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["pawelzajac95 <pawelzajac34@wp.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlalchemy = "^2.0.31"
psycopg2 = "^2.9.9"
alembic = "^1.13.2"
pydantic = "^2.7.4"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.1.3"
fastapi-mail = "^1.4.1"
redis = "^5.0.7"
asyncio-redis = "^0.16.0"
cloudinary = "^1.40.0"
fastapi-limiter = "^0.1.6"
sphinx = "^7.3.7"
pytest = "^8.2.2"
pytest-mock = "^3.14.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.3.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]

